In make-hunspell-dict, also generate en_GB hunspell dictionaries.
In addition update the README to make it more generic and make the
it's generation a script that gets the version, date, and level from
the environment.  If these are not provide in the environment sensible
defaults are used. The version defaults the SVN version if the working
copy was checked out from SVN.

[svn rev 183]

= a6cb416a3299ce0745fa006f43e7bca4586e8421