(Add a list of not so common contractions, but don't include them just yet.)
Alan wrote on issue #127: The contractions are all common in speech.
I must admit that mostly I've never seen these in writing, however.
How'd and how're are the only ones that seem remotely common in writing.
(But I don't think it would be a bad thing to add them all.)

NO CHANGE.
= c03538dc0376553e69559caec6a0a47b0f50ae7c