(libscowl: refactor)
Move tags into there own class for additional functionality.

Factor out most of the functionality of exportDB into _exportGroup so that it
is possible to export a single group at a time.

Add merge method to Groups and use it instead of an inner function of the
same name inside the LineBase.parse method.

NO CHANGE.
= 57dc7c9cc84cd8fe05900f474a646d70f524a1fb